+++ app/Http/Controllers/Site/SellerController.php
@@ -7,7 +7,6 @@ use App\Models\Admin\Marketplace;
 use App\Models\Site\Client;
 use App\Models\Site\Order;
 use App\Models\Site\Seller;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -19,30 +18,36 @@ class SellerController extends Controller
      */
     public function index()
     {
-        $sellers = Seller::withTrashed()->get();
+        $sellerModel = new Seller();
+
+        $sellers = $sellerModel->readAllSellers();
 
         return view('admin.sellers.index', compact('sellers'));
     }
 
     /**
      * Show one Seller's personal page.
+     *
+     * @param Request $request
      */
     public function show(Request $request)
     {
         $idSeller = $request->session()->get('id_seller');
+        $sellerModel = new Seller();
 
-        $seller = Seller::find($idSeller);
+        $seller = $sellerModel->readSellerWithCountry($idSeller);
 
         return view('profile.seller-show', compact('seller'));
     }
 
     /**
      * Display a listing of the Products from given Seller.
+     *
+     * @param Request $request
      */
     public function sellerProducts(Request $request)
     {
         $idSeller = $request->session()->get('id_seller');
-
         $seller = Seller::find($idSeller);
         $products = $seller->products;
 
@@ -51,11 +56,12 @@ class SellerController extends Controller
 
     /**
      * Display a listing of the Orders from given Seller.
+     *
+     * @param Request $request
      */
     public function sellerOrders(Request $request)
     {
         $idSeller = $request->session()->get('id_seller');
-
         $seller = Seller::find($idSeller);
         $orders = $seller->orders;
 
@@ -66,7 +72,6 @@ class SellerController extends Controller
 
             $idClient = $order->id_client;
             $client = Client::find($idClient);
-
             $order->client_name = $client->name;
             $order->client_surname = $client->surname;
             $order->client_email = $client->email;
@@ -78,15 +83,19 @@ class SellerController extends Controller
 
     /**
      * Update active Orders from given Seller.
+     *
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function sellerOrdersUpdate(Request $request)
+    public function sellerOrdersUpdate(Request $request): RedirectResponse
     {
         $idOrder = $request->post('id_order');
-
-        if ($request->post('order_accept')) {
-            Order::where('id_order', $idOrder)->update(['status' => 'processed']);
-        } elseif ($request->post('order_decline')) {
-            Order::where('id_order', $idOrder)->update(['status' => 'declined']);
+        if ($request->has('order_accept')) {
+            Order::where('id_order', $idOrder)
+                    ->update(['status' => 'processed']);
+        } elseif ($request->has('order_decline')) {
+            Order::where('id_order', $idOrder)
+                    ->update(['status' => 'declined']);
         }
 
         return redirect()->route('seller.my_orders');
@@ -99,8 +108,10 @@ class SellerController extends Controller
      */
     public function edit($idSeller)
     {
+        $sellerModel = new Seller();
+
+        $seller = $sellerModel->readSeller($idSeller);
         $marketplaces = Marketplace::all(['id_marketplace', 'country']);
-        $seller = Seller::find($idSeller);
 
         return view('profile.seller-update', compact('marketplaces', 'seller'));
     }
@@ -115,23 +126,24 @@ class SellerController extends Controller
     {
         $sellerModel = new Seller();
 
-        $postData = $request->post();
-        $setSellerData = [
-            'id_marketplace' => $postData['id_marketplace'],
-            'name' => $postData['name'],
-            'surname' => $postData['surname'],
-            'email' => $postData['email'],
-            'phone' => $postData['tel'],
-            'updated_at' => date('Y-m-d H:i:s'),
-        ];
+        if ($request->has('updateSeller')) {
+            $setSellerData = [
+                'id_marketplace' => $request->post('id_marketplace'),
+                'name' => $request->post('name'),
+                'surname' => $request->post('surname'),
+                'email' => $request->post('email'),
+                'phone' => $request->post('tel'),
+                'updated_at' => date('Y-m-d H:i:s'),
+            ];
+            $idSeller = $request->post('id_seller');
+            $sellerModel->updateSeller($idSeller, $setSellerData);
 
-        $idSeller = $request->post('id_seller');
-        $sellerModel->updateSeller($idSeller, $setSellerData);
-        $setSellerPasswordData = [
-            'password' => Hash::make($postData['password']),
-            'updated_at' => date('Y-m-d H:i:s'),
-        ];
-        $sellerModel->updateSellerPassword($idSeller, $setSellerPasswordData);
+            $setSellerPasswordData = [
+                'password' => Hash::make($request->post('password')),
+                'updated_at' => date('Y-m-d H:i:s'),
+            ];
+            $sellerModel->updateSellerPassword($idSeller, $setSellerPasswordData);
+        }
 
         return redirect()->route('seller.personal');
     }
@@ -144,11 +156,13 @@ class SellerController extends Controller
      */
     public function block(Request $request): RedirectResponse
     {
-        $idSeller = $request->post('id_seller');
-        $seller = Seller::find($idSeller);
-        $seller->delete();
+        if ($request->has('blockSeller')) {
+            $idSeller = $request->post('id_seller');
+            $seller = Seller::find($idSeller);
+            $seller->delete();
+        }
 
-        return redirect()->route('admin.seller');
+        return back();
     }
 
     /**
@@ -159,11 +173,13 @@ class SellerController extends Controller
      */
     public function unblock(Request $request): RedirectResponse
     {
-        $idSeller = $request->post('id_seller');
-        $seller = Seller::onlyTrashed()->find($idSeller);
-        $seller->restore();
+        if ($request->has('unblockSeller')) {
+            $idSeller = $request->post('id_seller');
+            $seller = Seller::onlyTrashed()->find($idSeller);
+            $seller->restore();
+        }
 
-        return redirect()->route('admin.seller');
+        return back();
     }
 
     /**
@@ -174,12 +190,13 @@ class SellerController extends Controller
      */
     public function destroy(Request $request): RedirectResponse
     {
-        $sellerModel = new Seller();
+        if ($request->has('deleteSeller')) {
+            $sellerModel = new Seller();
 
-        $idSeller = $request->post('id_seller');
-        $sellerModel->deleteSeller($idSeller);
-
-        $request->session()->forget('id_seller');
+            $idSeller = $request->post('id_seller');
+            $sellerModel->deleteSeller($idSeller);
+            $request->session()->forget('id_seller');
+        }
 
         return redirect()->route('index');
     }
